Fix password reset URL, add collab/battle/gifting system, harden forgot-password email
- Fix Brand:PrimaryResetUrl from /auth/reset-password to /reset-password (Next.js route groups don't appear in URLs — users were getting 404 from reset emails)
- Wrap forgot-password email send in try/catch so Graph API failures return 200 instead of resetting the TCP connection (caused 'Something went wrong' in browser)
- Add CollabController, CollabSession model, and migrations for collab sessions
- Add BattleAutoResolveService, BattleMatchmakingService, GiftBroadcastService
- Add artist slab split, decimal pieces, and battle score migrations
- Update GiftsController, WalletController, BattlesController for gifting system
- Add Brand:PrimaryResetUrl to appsettings.Production.json

+++ Models/Gifts/UserWallet.cs
@@ -14,8 +14,8 @@ public class UserWallet
     // Purchased currency — spent on gifts
     public int Slabs { get; set; } = 0;
 
-    // Earned back from gifting — 4 pieces = 1 slab purchasing power
-    public int Pieces { get; set; } = 0;
+    // Earned from gifts — 4 pieces = 1 slab; stored as decimal (battles award 1.5× e.g. 7.5)
+    public decimal Pieces { get; set; } = 0m;
 
     public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
 }
